feat: 상세 페이지 클라이언트 컴포넌트 추가 및 서버 컴포넌트로 변경
